Add support for plantweb module

+++ iplantuml/__init__.py
@@ -1,8 +1,12 @@
+import argparse
 import os
 import subprocess
+import uuid
 
 from IPython.core.magic import register_cell_magic
 from IPython.display import SVG
+
+import plantweb # dummy import to ensure plantweb module is present
 
 __title__ = "iplantuml"
 __description__ = "Package which adds a PlantUML cell magic to IPython."
@@ -10,7 +14,7 @@ __uri__ = "https://github.com/jbn/iplantuml"
 __doc__ = __description__ + " <" + __uri__ + ">"
 __license__ = "MIT"
 __copyright__ = "Copyright (c) 2017 John Bjorn Nelson"
-__version__ = "0.0.6"
+__version__ = "0.1.0"
 __author__ = "John Bjorn Nelson"
 __email__ = "jbn@abreka.com"
 
@@ -37,6 +41,22 @@ def plantuml_exec(*file_names, **kwargs):
 
     return [os.path.splitext(f)[0] + ".svg" for f in file_names]
 
+def plantuml_web(*file_names, **kwargs):
+    """
+    Given a list of UML documents, generate corresponding SVG diagrams, using
+    PlantUML's web service via the plantweb module.
+
+    :param file_names: the filenames of the documents for parsing by PlantUML.
+    :return: the path to the generated SVG UML diagram.
+    """
+
+    cmd = ["plantweb",
+           "--format",
+           "auto"] + list(file_names)
+
+    subprocess.check_call(cmd, shell=False, stderr=subprocess.STDOUT)
+
+    return [os.path.splitext(f)[0] + ".svg" for f in file_names]
 
 @register_cell_magic
 def plantuml(line, cell):
@@ -48,22 +68,27 @@ def plantuml(line, cell):
     :param cell: the PlantUML language UML specification.
     :return: a IPython SVG object for the diagram or None given error.
     """
-    if line:
-        retain, base_name = True, os.path.splitext(line)[0]
-    else:
-        retain, base_name = False, "tmp"
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-j", "--jar", action="store_true", help="render using plantuml.jar (default is web service)")
+    parser.add_argument("-n", "--name",  type=str, default=None, help="persist as <name>.uml and <name>.svg after rendering")
+    parser.add_argument("-p", "--plantuml-path", default=None, help="specify PlantUML jar path (default=%s)" % PLANTUMLPATH)
+    args = parser.parse_args(line.split() if line else "")
+    retain = args.name is not None
+    base_name = args.name or str(uuid.uuid4())
+    use_web = not (args.jar or args.plantuml_path)
 
     uml_path = base_name + ".uml"
     with open(uml_path, 'w') as fp:
         fp.write(cell)
 
     try:
-        svg_name = plantuml_exec(uml_path)[0]
+        output = plantuml_web(uml_path) if use_web else plantuml_exec(uml_path, plantuml_path=os.path.abspath(args.plantuml_path or PLANTUMLPATH))
+        svg_name = output[0]
         return SVG(filename=svg_name)
     finally:
         if not retain:
-            os.unlink(uml_path)
-
+            if os.path.exists(uml_path): os.unlink(uml_path)
             svg_path = base_name + ".svg"
-            if os.path.exists(svg_path):
-                os.unlink(svg_path)
+            if os.path.exists(svg_path): os.unlink(svg_path)
+
